fin
ds

+++ Uniconekt_v2.0/app/src/main/java/com/dwmedios/uniconekt/view/util/Utils.java
@@ -15,7 +15,6 @@ import android.support.v4.content.ContextCompat;
 import android.support.v7.app.ActionBar;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
-import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
 import android.view.animation.AnimationUtils;
@@ -359,9 +358,10 @@ public class Utils {
                                 succ = folder.mkdirs();
                             } else {
                                 File[] files = folder.listFiles();
-                                for (File mFile : files) {
-                                    mFile.delete();
-                                }
+                                if (files != null)
+                                    for (File mFile : files) {
+                                        mFile.delete();
+                                    }
                             }
 
                             if (succ) {
